+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 * VPC module
 
+### Changed
+* Launch configuration default ami to match the one in us-east-1
+
 ## 0.0.1 - 2016-12-24
 ### Added
 * Launch configuration module